Move `GroupedProductListSelectorDataSource` to a separate file.
